+++ lib/src/models/notification_data.dart
@@ -4,7 +4,7 @@ part 'notification_data.freezed.dart';
 part 'notification_data.g.dart';
 
 @freezed
-class NotificationData with _$NotificationData {
+abstract class NotificationData with _$NotificationData {
   const factory NotificationData({
     required String title,
     required String body,