+++ lib/src/models/notification_data.freezed.dart
@@ -1,5 +1,5 @@
-// coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// coverage:ignore-file
 // ignore_for_file: type=lint
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
 
@@ -9,235 +9,278 @@ part of 'notification_data.dart';
 // FreezedGenerator
 // **************************************************************************
 
+// dart format off
 T _$identity<T>(T value) => value;
-
-final _privateConstructorUsedError = UnsupportedError(
-  'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models',
-);
-
-NotificationData _$NotificationDataFromJson(Map<String, dynamic> json) {
-  return _NotificationData.fromJson(json);
-}
 
 /// @nodoc
 mixin _$NotificationData {
-  String get title => throw _privateConstructorUsedError;
-  String get body => throw _privateConstructorUsedError;
-  String? get payload => throw _privateConstructorUsedError;
-  NotificationType get type => throw _privateConstructorUsedError;
+
+ String get title; String get body; String? get payload; NotificationType get type;
+/// Create a copy of NotificationData
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$NotificationDataCopyWith<NotificationData> get copyWith => _$NotificationDataCopyWithImpl<NotificationData>(this as NotificationData, _$identity);
 
   /// Serializes this NotificationData to a JSON map.
-  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+  Map<String, dynamic> toJson();
 
-  /// Create a copy of NotificationData
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  $NotificationDataCopyWith<NotificationData> get copyWith =>
-      throw _privateConstructorUsedError;
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is NotificationData&&(identical(other.title, title) || other.title == title)&&(identical(other.body, body) || other.body == body)&&(identical(other.payload, payload) || other.payload == payload)&&(identical(other.type, type) || other.type == type));
+}
+
+@JsonKey(includeFromJson: false, includeToJson: false)
+@override
+int get hashCode => Object.hash(runtimeType,title,body,payload,type);
+
+@override
+String toString() {
+  return 'NotificationData(title: $title, body: $body, payload: $payload, type: $type)';
+}
+
+
 }
 
 /// @nodoc
-abstract class $NotificationDataCopyWith<$Res> {
-  factory $NotificationDataCopyWith(
-    NotificationData value,
-    $Res Function(NotificationData) then,
-  ) = _$NotificationDataCopyWithImpl<$Res, NotificationData>;
-  @useResult
-  $Res call({
-    String title,
-    String body,
-    String? payload,
-    NotificationType type,
-  });
-}
+abstract mixin class $NotificationDataCopyWith<$Res>  {
+  factory $NotificationDataCopyWith(NotificationData value, $Res Function(NotificationData) _then) = _$NotificationDataCopyWithImpl;
+@useResult
+$Res call({
+ String title, String body, String? payload, NotificationType type
+});
 
+
+
+
+}
 /// @nodoc
-class _$NotificationDataCopyWithImpl<$Res, $Val extends NotificationData>
+class _$NotificationDataCopyWithImpl<$Res>
     implements $NotificationDataCopyWith<$Res> {
-  _$NotificationDataCopyWithImpl(this._value, this._then);
+  _$NotificationDataCopyWithImpl(this._self, this._then);
 
-  // ignore: unused_field
-  final $Val _value;
-  // ignore: unused_field
-  final $Res Function($Val) _then;
+  final NotificationData _self;
+  final $Res Function(NotificationData) _then;
 
-  /// Create a copy of NotificationData
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? title = null,
-    Object? body = null,
-    Object? payload = freezed,
-    Object? type = null,
-  }) {
-    return _then(
-      _value.copyWith(
-            title: null == title
-                ? _value.title
-                : title // ignore: cast_nullable_to_non_nullable
-                      as String,
-            body: null == body
-                ? _value.body
-                : body // ignore: cast_nullable_to_non_nullable
-                      as String,
-            payload: freezed == payload
-                ? _value.payload
-                : payload // ignore: cast_nullable_to_non_nullable
-                      as String?,
-            type: null == type
-                ? _value.type
-                : type // ignore: cast_nullable_to_non_nullable
-                      as NotificationType,
-          )
-          as $Val,
-    );
-  }
+/// Create a copy of NotificationData
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? title = null,Object? body = null,Object? payload = freezed,Object? type = null,}) {
+  return _then(_self.copyWith(
+title: null == title ? _self.title : title // ignore: cast_nullable_to_non_nullable
+as String,body: null == body ? _self.body : body // ignore: cast_nullable_to_non_nullable
+as String,payload: freezed == payload ? _self.payload : payload // ignore: cast_nullable_to_non_nullable
+as String?,type: null == type ? _self.type : type // ignore: cast_nullable_to_non_nullable
+as NotificationType,
+  ));
 }
 
-/// @nodoc
-abstract class _$$NotificationDataImplCopyWith<$Res>
-    implements $NotificationDataCopyWith<$Res> {
-  factory _$$NotificationDataImplCopyWith(
-    _$NotificationDataImpl value,
-    $Res Function(_$NotificationDataImpl) then,
-  ) = __$$NotificationDataImplCopyWithImpl<$Res>;
-  @override
-  @useResult
-  $Res call({
-    String title,
-    String body,
-    String? payload,
-    NotificationType type,
-  });
 }
 
-/// @nodoc
-class __$$NotificationDataImplCopyWithImpl<$Res>
-    extends _$NotificationDataCopyWithImpl<$Res, _$NotificationDataImpl>
-    implements _$$NotificationDataImplCopyWith<$Res> {
-  __$$NotificationDataImplCopyWithImpl(
-    _$NotificationDataImpl _value,
-    $Res Function(_$NotificationDataImpl) _then,
-  ) : super(_value, _then);
 
-  /// Create a copy of NotificationData
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? title = null,
-    Object? body = null,
-    Object? payload = freezed,
-    Object? type = null,
-  }) {
-    return _then(
-      _$NotificationDataImpl(
-        title: null == title
-            ? _value.title
-            : title // ignore: cast_nullable_to_non_nullable
-                  as String,
-        body: null == body
-            ? _value.body
-            : body // ignore: cast_nullable_to_non_nullable
-                  as String,
-        payload: freezed == payload
-            ? _value.payload
-            : payload // ignore: cast_nullable_to_non_nullable
-                  as String?,
-        type: null == type
-            ? _value.type
-            : type // ignore: cast_nullable_to_non_nullable
-                  as NotificationType,
-      ),
-    );
-  }
+/// Adds pattern-matching-related methods to [NotificationData].
+extension NotificationDataPatterns on NotificationData {
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
+
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>(TResult Function( _NotificationData value)?  $default,{required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _NotificationData() when $default != null:
+return $default(_that);case _:
+  return orElse();
+
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
+
+@optionalTypeArgs TResult map<TResult extends Object?>(TResult Function( _NotificationData value)  $default,){
+final _that = this;
+switch (_that) {
+case _NotificationData():
+return $default(_that);case _:
+  throw StateError('Unexpected subclass');
+
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
+
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(TResult? Function( _NotificationData value)?  $default,){
+final _that = this;
+switch (_that) {
+case _NotificationData() when $default != null:
+return $default(_that);case _:
+  return null;
+
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
+
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( String title,  String body,  String? payload,  NotificationType type)?  $default,{required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _NotificationData() when $default != null:
+return $default(_that.title,_that.body,_that.payload,_that.type);case _:
+  return orElse();
+
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
+
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( String title,  String body,  String? payload,  NotificationType type)  $default,) {final _that = this;
+switch (_that) {
+case _NotificationData():
+return $default(_that.title,_that.body,_that.payload,_that.type);case _:
+  throw StateError('Unexpected subclass');
+
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
+
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( String title,  String body,  String? payload,  NotificationType type)?  $default,) {final _that = this;
+switch (_that) {
+case _NotificationData() when $default != null:
+return $default(_that.title,_that.body,_that.payload,_that.type);case _:
+  return null;
+
+}
+}
+
 }
 
 /// @nodoc
 @JsonSerializable()
-class _$NotificationDataImpl implements _NotificationData {
-  const _$NotificationDataImpl({
-    required this.title,
-    required this.body,
-    this.payload,
-    this.type = NotificationType.info,
-  });
 
-  factory _$NotificationDataImpl.fromJson(Map<String, dynamic> json) =>
-      _$$NotificationDataImplFromJson(json);
+class _NotificationData implements NotificationData {
+  const _NotificationData({required this.title, required this.body, this.payload, this.type = NotificationType.info});
+  factory _NotificationData.fromJson(Map<String, dynamic> json) => _$NotificationDataFromJson(json);
 
-  @override
-  final String title;
-  @override
-  final String body;
-  @override
-  final String? payload;
-  @override
-  @JsonKey()
-  final NotificationType type;
+@override final  String title;
+@override final  String body;
+@override final  String? payload;
+@override@JsonKey() final  NotificationType type;
 
-  @override
-  String toString() {
-    return 'NotificationData(title: $title, body: $body, payload: $payload, type: $type)';
-  }
+/// Create a copy of NotificationData
+/// with the given fields replaced by the non-null parameter values.
+@override @JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+_$NotificationDataCopyWith<_NotificationData> get copyWith => __$NotificationDataCopyWithImpl<_NotificationData>(this, _$identity);
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$NotificationDataImpl &&
-            (identical(other.title, title) || other.title == title) &&
-            (identical(other.body, body) || other.body == body) &&
-            (identical(other.payload, payload) || other.payload == payload) &&
-            (identical(other.type, type) || other.type == type));
-  }
-
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  int get hashCode => Object.hash(runtimeType, title, body, payload, type);
-
-  /// Create a copy of NotificationData
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$NotificationDataImplCopyWith<_$NotificationDataImpl> get copyWith =>
-      __$$NotificationDataImplCopyWithImpl<_$NotificationDataImpl>(
-        this,
-        _$identity,
-      );
-
-  @override
-  Map<String, dynamic> toJson() {
-    return _$$NotificationDataImplToJson(this);
-  }
+@override
+Map<String, dynamic> toJson() {
+  return _$NotificationDataToJson(this, );
 }
 
-abstract class _NotificationData implements NotificationData {
-  const factory _NotificationData({
-    required final String title,
-    required final String body,
-    final String? payload,
-    final NotificationType type,
-  }) = _$NotificationDataImpl;
-
-  factory _NotificationData.fromJson(Map<String, dynamic> json) =
-      _$NotificationDataImpl.fromJson;
-
-  @override
-  String get title;
-  @override
-  String get body;
-  @override
-  String? get payload;
-  @override
-  NotificationType get type;
-
-  /// Create a copy of NotificationData
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  _$$NotificationDataImplCopyWith<_$NotificationDataImpl> get copyWith =>
-      throw _privateConstructorUsedError;
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _NotificationData&&(identical(other.title, title) || other.title == title)&&(identical(other.body, body) || other.body == body)&&(identical(other.payload, payload) || other.payload == payload)&&(identical(other.type, type) || other.type == type));
 }
+
+@JsonKey(includeFromJson: false, includeToJson: false)
+@override
+int get hashCode => Object.hash(runtimeType,title,body,payload,type);
+
+@override
+String toString() {
+  return 'NotificationData(title: $title, body: $body, payload: $payload, type: $type)';
+}
+
+
+}
+
+/// @nodoc
+abstract mixin class _$NotificationDataCopyWith<$Res> implements $NotificationDataCopyWith<$Res> {
+  factory _$NotificationDataCopyWith(_NotificationData value, $Res Function(_NotificationData) _then) = __$NotificationDataCopyWithImpl;
+@override @useResult
+$Res call({
+ String title, String body, String? payload, NotificationType type
+});
+
+
+
+
+}
+/// @nodoc
+class __$NotificationDataCopyWithImpl<$Res>
+    implements _$NotificationDataCopyWith<$Res> {
+  __$NotificationDataCopyWithImpl(this._self, this._then);
+
+  final _NotificationData _self;
+  final $Res Function(_NotificationData) _then;
+
+/// Create a copy of NotificationData
+/// with the given fields replaced by the non-null parameter values.
+@override @pragma('vm:prefer-inline') $Res call({Object? title = null,Object? body = null,Object? payload = freezed,Object? type = null,}) {
+  return _then(_NotificationData(
+title: null == title ? _self.title : title // ignore: cast_nullable_to_non_nullable
+as String,body: null == body ? _self.body : body // ignore: cast_nullable_to_non_nullable
+as String,payload: freezed == payload ? _self.payload : payload // ignore: cast_nullable_to_non_nullable
+as String?,type: null == type ? _self.type : type // ignore: cast_nullable_to_non_nullable
+as NotificationType,
+  ));
+}
+
+
+}
+
+// dart format on