+++ lib/src/models/notification_data.g.dart
@@ -6,25 +6,23 @@ part of 'notification_data.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$NotificationDataImpl _$$NotificationDataImplFromJson(
-  Map<String, dynamic> json,
-) => _$NotificationDataImpl(
-  title: json['title'] as String,
-  body: json['body'] as String,
-  payload: json['payload'] as String?,
-  type:
-      $enumDecodeNullable(_$NotificationTypeEnumMap, json['type']) ??
-      NotificationType.info,
-);
+_NotificationData _$NotificationDataFromJson(Map<String, dynamic> json) =>
+    _NotificationData(
+      title: json['title'] as String,
+      body: json['body'] as String,
+      payload: json['payload'] as String?,
+      type:
+          $enumDecodeNullable(_$NotificationTypeEnumMap, json['type']) ??
+          NotificationType.info,
+    );
 
-Map<String, dynamic> _$$NotificationDataImplToJson(
-  _$NotificationDataImpl instance,
-) => <String, dynamic>{
-  'title': instance.title,
-  'body': instance.body,
-  'payload': instance.payload,
-  'type': _$NotificationTypeEnumMap[instance.type]!,
-};
+Map<String, dynamic> _$NotificationDataToJson(_NotificationData instance) =>
+    <String, dynamic>{
+      'title': instance.title,
+      'body': instance.body,
+      'payload': instance.payload,
+      'type': _$NotificationTypeEnumMap[instance.type]!,
+    };
 
 const _$NotificationTypeEnumMap = {
   NotificationType.info: 'info',